refactoring, fixes, and initial tests
