+++ mist-lib/src/services/St1Service/StpiMatchHelper.js
@@ -1,0 +1,184 @@
+'use strict'
+
+// Vendor
+const _ = require('lodash')
+
+// Local
+const { pfamHDUnrelatedSignalDomains, pfamHKNonSignalDomains } = require('./stpi.constants')
+
+function setContainsSomeDomains(someSet, domains) {
+  if (!someSet || !domains)
+    return false
+
+  for (let domain of domains) {
+    if (someSet.contains(domain.name))
+      return true
+  }
+  return false
+}
+
+module.exports =
+class StpiMatchHelper {
+  constructor(stpiSpec) {
+    const filters = {
+      isMarker: {marker: true},
+      isHK_CA: {id: 'HK_CA'},
+      isHisKA: {id: 'HisKA'},
+      isChemotaxis: {kind: 'chemotaxis'},
+      isTransmitter: {kind: 'transmitter'},
+      isReceiver: {kind: 'receiver'},
+      isOutput: {kind: 'output'},
+    }
+
+    const agfam = _.filter(stpiSpec, {family: 'agfam'})
+    const ecf = _.filter(stpiSpec, {family: 'ecf'})
+    const pfam = _.filter(stpiSpec, {family: 'pfam'})
+    const pfamEcf = _.filter(pfam, {kind: 'ecf'})
+    this.groups = {
+      agfam,
+      agfamMarker: _.filter(agfam, filters.isMarker),
+      agfamHatpase: _.filter(agfam, filters.isHK_CA),
+      agfamTransmitter: _.filter(agfam, filters.isTransmitter),
+      agfamReceiver: _.filter(agfam, filters.isReceiver),
+      agfamChemo: _.filter(agfam, filters.isChemotaxis),
+      agfamOutput: _.filter(agfam, filters.isOutput),
+
+      ecf,
+      ecfMarker: _.filter(ecf, filters.isMarker),
+
+      pfam,
+      pfamMarker: _.filter(pfam, filters.isMarker),
+      pfamHatpase: _.filter(pfam, filters.isHK_CA),
+      pfamHiska: _.filter(pfam, filters.isHisKA),
+      pfamChemo: _.filter(pfam, filters.isChemotaxis),
+      pfamTransmitter: _.filter(pfam, filters.isTransmitter),
+      pfamReceiver: _.filter(pfam, filters.isReceiver),
+      pfamEcf,
+      pfamEcfMarker: _.filter(pfamEcf, filters.isMarker),
+      pfamOutput: _.filter(pfam, filters.isOutput),
+    }
+
+    // Create sets of unique domain names for each group for quickly determining
+    // membership to a group
+    // e.g. [agfam] = Set([RR, HK_CA, HK_CA:1, ...])
+    this.sets = {
+      // Deprecated: 9 Sep 2018
+      hdUnrelatedSignalDomains: pfamHDUnrelatedSignalDomains,
+      hkNonSignalDomains: pfamHKNonSignalDomains,
+    }
+
+    //      {group name}->{domain name} = signal domain
+    // e.g. {agfam}->{RR} = {family: agfam, name: RR, ...}
+    this.toSignalDomain = {}
+    for (let [groupName, signalDomains] of Object.entries(this.groups)) {
+      const nameToSignalDomain = {}
+      signalDomains.forEach((signalDomain) => {
+        nameToSignalDomain[signalDomain.name] = signalDomain
+      })
+      this.toSignalDomain[groupName] = nameToSignalDomain
+
+      this.sets[groupName] = new Set(signalDomains.map((signalDomain) => signalDomain.name))
+    }
+
+    this.signalDomainIdToPfamNames = {}
+    this.groups.pfam.forEach((signalDomain) => {
+      const { id, name } = signalDomain
+      const set = this.signalDomainIdToPfamNames[id]
+      if (!set) {
+        set = this.signalDomainIdToPfamNames[id] = new Set()
+      }
+      set.add(name)
+    })
+  }
+
+  contains(domains, name) {
+    if (!domains)
+      return false
+
+    for (let i = 0, z = domains.length; i < z; ++i) {
+      if (domains[i].name === name)
+        return true
+    }
+
+    return false
+  }
+
+  hasPrimaryMarker(bundle) {
+    return setContainsSomeDomains(bundle.pfam, this.sets.pfamMarker) ||
+      setContainsSomeDomains(bundle.agfam, this.sets.agfamMarker)
+  }
+
+  hasEcfMarker(bundle) {
+    return setContainsSomeDomains(bundle.ecf, this.sets.ecfMarker)
+  }
+
+  hasHatpase(bundle) {
+    return setContainsSomeDomains(bundle.pfam, this.sets.pfamHatpase) ||
+      setContainsSomeDomains(bundle.pfam, this.sets.pfamHiska) ||
+      setContainsSomeDomains(bundle.agfam, this.sets.agfamHatpase)
+  }
+
+  hasHpt(bundle) {
+    return this.contains(pfam, 'Hpt')
+  }
+
+  hasTransmitter(bundle) {
+    return setContainsSomeDomains(bundle.pfam, this.sets.pfamTransmitter) ||
+      setContainsSomeDomains(bundle.agfam, this.sets.agfamTransmitter)
+  }
+
+  hasReceiver(bundle) {
+    return setContainsSomeDomains(bundle.pfam, this.sets.pfamReceiver) ||
+      setContainsSomeDomains(bundle.agfam, this.sets.agfamReceiver)
+  }
+
+  hasOutput(bundle) {
+    return setContainsSomeDomains(bundle.agfam, this.sets.agfamOutput) ||
+      setContainsSomeDomains(bundle.pfam, this.sets.pfamOutput)
+  }
+
+  hasPfamEcf(bundle) {
+    return setContainsSomeDomains(bundle.pfam, this.sets.pfamEcf)
+  }
+
+  hasPfamEcfMarker(bundle) {
+    return setContainsSomeDomains(bundle.pfam, this.sets.pfamEcfMarker)
+  }
+
+  isChemotaxis(bundle) {
+    return setContainsSomeDomains(bundle.pfam, this.sets.pfamChemo) ||
+      setContainsSomeDomains(bundle.afam, this.sets.agfamChemo)
+  }
+
+  // Deprecated: 9 Sep 2018
+  isHdWithUnrelatedDomains(bundle) {
+    return this.contains(bundle.pfam, 'HD') &&
+      setContainsSomeDomains(bundle.pfam, this.sets.hdUnrelatedSignalDomains)
+  }
+
+  isNonSignalingHK(bundle) {
+    return this.hasHatpase(bundle) &&
+      setContainsSomeDomains(bundle.pfam, this.sets.hkNonSignalDomains)
+  }
+
+  findNterminalHatpase(bundle) {
+    let hatpase = this.findNterminalKind_(bundle.agfam, this.sets.agfamHatpase)
+    hatpase = this.findNterminalHatpase(bundle.pfam, this.sets.pfamHatpase, hatpase)
+    return hatpase
+  }
+
+  findNterminalReceiver(bundle) {
+    let receiver = this.findNterminalKind_(bundle.agfam, this.sets.agfamReceiver)
+    receiver = this.findNterminalHatpase(bundle.pfam, this.sets.pfamReceiver, receiver)
+    return receiver
+  }
+
+  findNterminalKind_(domains, targetSet, currentMatch = null) {
+    for (let domain of domains) {
+      if (targetSet.contains(domain.name) && (!currentMatch || domain.start < currentMatch.start)) {
+        currentMatch = domain
+      }
+    }
+    return currentMatch
+  }
+}
